requirements.txt
anyio==3.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
exceptiongroup==1.1.1
fastapi==0.97.0
google==3.0.0
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
h11==0.14.0
httptools==0.5.0
idna==3.4
paramiko==3.2.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyMySQL==1.0.3
PyNaCl==1.5.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
starlette==0.27.0
storage==0.0.4.3
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3

pandas~=2.0.2
redis~=4.5.5
boto3~=1.26.154
botocore~=1.29.154
locust~=2.15.1